streamlit/streamlit/requirements.txt
functions_framework
Flask
pandas
numpy
google-cloud-bigquery
google-cloud-storage
google-auth
google-auth-oauthlib
streamlit
sqlparse
googletrans==4.0.0-rc1
langfuse
openai
vacances_scolaires_france==0.10.0
plotly==5.23.0
prophet
scipy==1.14.0